+++ CHANGELOG.md
@@ -37,6 +37,7 @@
  * Add schema validation plugin to validate schema during compilation time
  * Add support to ignore implemented interfaces on objects and child interfaces
  * Add support to load symfony4 data fixtures in path 'src/DataFixtures'
+ * Add command to export schema to standard output or file, as json or graphql format
  
 1.0.6 - 18-03-23
 ----